Address more feedback

+++ posixlib/src/main/scala/scala/scalanative/runtime/SocketHelpers.scala
@@ -77,7 +77,7 @@ object SocketHelpers {
         }
 
         val sentBytes = send(sock, toCString("echo"), 4, 0)
-        if (sentBytes.toLong < 4) {
+        if (sentBytes < 4) {
           return false
         }
 
@@ -92,7 +92,7 @@ object SocketHelpers {
         } else {
           val buf      = alloc[CChar](5)
           val recBytes = recv(sock, buf, 5, 0)
-          if (recBytes.toLong < 4) {
+          if (recBytes < 4) {
             return false
           }
         }
